Use uint types in itm-impl tests
Migrate tests to use proper types, eliminating associated warnings.

+++ itm/itm-impl/src/test/java/org/opendaylight/genius/itm/tests/OvsdbTestUtil.java
@@ -38,6 +38,7 @@ import org.opendaylight.yang.gen.v1.urn.tbd.params.xml.ns.yang.network.topology.
 import org.opendaylight.yang.gen.v1.urn.tbd.params.xml.ns.yang.network.topology.rev131021.network.topology.topology.NodeBuilder;
 import org.opendaylight.yang.gen.v1.urn.tbd.params.xml.ns.yang.network.topology.rev131021.network.topology.topology.NodeKey;
 import org.opendaylight.yangtools.yang.binding.InstanceIdentifier;
+import org.opendaylight.yangtools.yang.common.Uint16;
 
 public final class OvsdbTestUtil {
     private OvsdbTestUtil() {
@@ -47,7 +48,7 @@ public final class OvsdbTestUtil {
     /* methods */
     public static ConnectionInfo getConnectionInfo(int port, String strIpAddress) {
         IpAddress ipAddress = IpAddressBuilder.getDefaultInstance(strIpAddress);
-        PortNumber portNumber = new PortNumber(port);
+        PortNumber portNumber = new PortNumber(Uint16.valueOf(port));
 
         ConnectionInfo connectionInfo =
             new ConnectionInfoBuilder().setRemoteIp(ipAddress).setRemotePort(portNumber).build();
